+++ firebase-firestore/src/main/java/com/google/firebase/firestore/pipeline/stage.kt
@@ -14,7 +14,6 @@
 
 package com.google.firebase.firestore.pipeline
 
-import com.google.common.annotations.Beta
 import com.google.firebase.firestore.UserDataReader
 import com.google.firebase.firestore.VectorValue
 import com.google.firebase.firestore.model.Document
@@ -31,7 +30,6 @@ import com.google.firestore.v1.Pipeline
 import com.google.firestore.v1.Value
 import javax.annotation.Nonnull
 
-@Beta
 sealed class Stage<T : Stage<T>>(internal val name: String, internal val options: InternalOptions) {
   internal fun toProtoStage(userDataReader: UserDataReader): Pipeline.Stage {
     val builder = Pipeline.Stage.newBuilder()
@@ -110,7 +108,6 @@ sealed class Stage<T : Stage<T>>(internal val name: String, internal val options
  * This class provides a way to call stages that are supported by the Firestore backend but that are
  * not implemented in the SDK version being used.
  */
-@Beta
 class RawStage
 private constructor(
   name: String,
@@ -205,7 +202,6 @@ internal constructor(options: InternalOptions = InternalOptions.EMPTY) :
   }
 }
 
-@Beta
 class CollectionSource
 internal constructor(
   internal val path: ResourcePath,
@@ -253,7 +249,6 @@ internal constructor(
   }
 }
 
-@Beta
 class CollectionSourceOptions internal constructor(options: InternalOptions) :
   AbstractOptions<CollectionSourceOptions>(options) {
   /** Creates a new, empty `CollectionSourceOptions` object. */
@@ -272,7 +267,6 @@ class CollectionSourceOptions internal constructor(options: InternalOptions) :
   }
 }
 
-@Beta
 class CollectionHints internal constructor(options: InternalOptions) :
   AbstractOptions<CollectionHints>(options) {
   /** Creates a new, empty `CollectionHints` object. */
@@ -303,7 +297,6 @@ class CollectionHints internal constructor(options: InternalOptions) :
   }
 }
 
-@Beta
 class CollectionGroupSource
 internal constructor(val collectionId: String, options: InternalOptions) :
   Stage<CollectionGroupSource>("collection_group", options) {
@@ -344,7 +337,6 @@ internal constructor(val collectionId: String, options: InternalOptions) :
   }
 }
 
-@Beta
 class CollectionGroupOptions internal constructor(options: InternalOptions) :
   AbstractOptions<CollectionGroupOptions>(options) {
   /** Creates a new, empty `CollectionGroupOptions` object. */
@@ -473,7 +465,6 @@ internal constructor(
  * [AggregateFunction.alias] on [AggregateFunction] instances. Each aggregation calculates a value
  * (e.g., sum, average, count) based on the documents within its group.
  */
-@Beta
 class AggregateStage
 private constructor(
   private val accumulators: Map<String, AggregateFunction>,
@@ -574,7 +565,6 @@ private constructor(
   }
 }
 
-@Beta
 class AggregateHints internal constructor(options: InternalOptions) :
   AbstractOptions<AggregateHints>(options) {
   /** Creates a new, empty `AggregateHints` object. */
@@ -589,7 +579,6 @@ class AggregateHints internal constructor(options: InternalOptions) :
   }
 }
 
-@Beta
 class AggregateOptions internal constructor(options: InternalOptions) :
   AbstractOptions<AggregateOptions>(options) {
   /** Creates a new, empty `AggregateOptions` object. */
@@ -648,7 +637,6 @@ internal constructor(
  * Performs a vector similarity search, ordering the result set by most similar to least similar,
  * and returning the first N documents in the result set.
  */
-@Beta
 class FindNearestStage
 internal constructor(
   private val property: Expression,
@@ -795,7 +783,6 @@ internal constructor(
   }
 }
 
-@Beta
 class FindNearestOptions private constructor(options: InternalOptions) :
   AbstractOptions<FindNearestOptions>(options) {
   /** Creates a new, empty `FindNearestOptions` object. */
@@ -1128,7 +1115,6 @@ internal constructor(
  * dictate how the sample is calculated either by specifying a target output size, or by specifying
  * a target percentage of the input size.
  */
-@Beta
 class SampleStage
 private constructor(
   private val size: Number,
@@ -1228,7 +1214,6 @@ internal constructor(
  * Takes a specified array from the input documents and outputs a document for each element with the
  * element stored in a field with name specified by the alias.
  */
-@Beta
 class UnnestStage
 internal constructor(
   private val selectable: Selectable,
@@ -1302,7 +1287,6 @@ internal constructor(
   fun withIndexField(indexField: String): UnnestStage = withOption("index_field", indexField)
 }
 
-@Beta
 class UnnestOptions private constructor(options: InternalOptions) :
   AbstractOptions<UnnestOptions>(options) {
   /** Creates a new, empty `UnnestOptions` object. */